+++ src/main/java/org/ideoholic/curium/model/student/dto/StudentDto.java
@@ -54,7 +54,7 @@ public class StudentDto implements java.io.Serializable {
 
     private String placeofbirth;
 
-    private Integer nooftc;
+    private Long nooftc;
 
     private String dateoftc;
